requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
pydantic-settings==2.*
sqlalchemy==2.*
aiosqlite==0.20.*
httpx==0.27.*
langgraph==0.2.*
openai==1.*
python-dotenv==1.*
matplotlib==3.8.*
seaborn==0.13.*
pandas==2.1.*
plotly==5.17.*
kaleido==0.2.*
urllib3<2.0

